Refactored the backend codebase to make it more modular, modified contributing.md with better descriptions, improved README.
