+++ pyproject.toml
@@ -25,6 +25,7 @@ dependencies = [
     "beautifulsoup4>=4.12.0",
     "pandas>=2.0.0",
     "lxml>=4.9.0",
+    "pytest>=8.4.2",
 ]
 
 [project.optional-dependencies]
