+++ schemas/game_detection.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+from typing import Optional, List
+
+class GameDetectionRequest(BaseModel):
+    message: Optional[str] = None
