+++ schemas/settings.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+from typing import Optional, List
+class ApiKeyRequest(BaseModel):
+    api_key: str
