+++ services/game_detection.py
